+++ LibSm64Sharp/src/ISm64Context.cs
@@ -1,6 +1,6 @@
 ﻿namespace libsm64sharp {
   public interface ISm64Context : IDisposable {
-    ISm64Mario CreateMario(short x, short y, short z);
+    ISm64Mario CreateMario(float x, float y, float z);
 
     // TODO: Are builders the right design for this?
     ISm64DynamicCollisionMeshBuilder CreateDynamicCollisionMesh();
