+++ LibSm64Sharp/src/Sm64Triangle.cs
@@ -3,7 +3,7 @@
     private class Sm64Triangle : ISm64Triangle {
       public Sm64SurfaceType SurfaceType { get; init; }
       public Sm64TerrainType TerrainType { get; init; }
-      public IReadOnlyList<IReadOnlySm64Vector3<short>> Vertices { get; init; }
+      public IReadOnlyList<IReadOnlySm64Vector3<int>> Vertices { get; init; }
     }
   }
 }
